+++ app/login/input.tsx
@@ -1,0 +1,25 @@
+import { ChangeEvent } from "react";
+
+interface InputProps {
+  type: string;
+  placeholder: string;
+  value: string;
+  onChange: (e: ChangeEvent<HTMLInputElement>) => void;
+}
+
+const Input: React.FC<InputProps> = ({
+  type,
+  placeholder,
+  value,
+  onChange,
+}) => (
+  <input
+    type={type}
+    placeholder={placeholder}
+    value={value}
+    onChange={onChange}
+    className="w-full p-2 rounded-md border border-gray-300 mb-3"
+  />
+);
+
+export default Input;
